+++ src/internal/api/auth/infrastructure/handlers/update_user_handler_test.go
@@ -52,7 +52,7 @@ func TestUpdateUserHandler_Returns_An_Error_If_The_Query_To_Find_The_User_Fails(
 	}
 
 	req := request()
-	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: int32(1)}).Return(nil, fmt.Errorf("some error")).Once()
+	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: 1}).Return(nil, fmt.Errorf("some error")).Once()
 
 	result := UpdateUserHandler(httptest.NewRecorder(), req, h)
 
@@ -81,7 +81,7 @@ func TestUpdateUserHandler_Returns_An_ErrorResult_With_A_BadRequestError_If_Trie
 	req := request()
 	foundUserName, _ := domain.NewUserNameValueObject("admin")
 	foundUser := domain.UserEntity{Name: foundUserName}
-	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: int32(1)}).Return(&foundUser, nil).Once()
+	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: 1}).Return(&foundUser, nil).Once()
 
 	result := UpdateUserHandler(httptest.NewRecorder(), req, h)
 
@@ -110,7 +110,7 @@ func TestUpdateUserHandler_Returns_An_ErrorResult_With_A_BadRequestError_If_Trie
 	req := request()
 	foundUserName, _ := domain.NewUserNameValueObject("admin")
 	foundUser := domain.UserEntity{Name: foundUserName}
-	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: int32(1)}).Return(&foundUser, nil).Once()
+	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: 1}).Return(&foundUser, nil).Once()
 
 	result := UpdateUserHandler(httptest.NewRecorder(), req, h)
 
@@ -139,7 +139,7 @@ func TestUpdateUserHandler_Returns_An_Error_If_The_Query_To_Check_If_A_User_With
 	req := request()
 	foundUserName, _ := domain.NewUserNameValueObject("wadus")
 	foundUser := domain.UserEntity{Name: foundUserName}
-	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: int32(1)}).Return(&foundUser, nil).Once()
+	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: 1}).Return(&foundUser, nil).Once()
 	mockedUsersRepo.On("ExistsUser", req.Context(), domain.UserEntity{Name: userName}).Return(false, fmt.Errorf("some error")).Once()
 
 	result := UpdateUserHandler(httptest.NewRecorder(), req, h)
@@ -170,7 +170,7 @@ func TestUpdateUserHandler_Returns_An_ErrorResult_With_A_BadRequestError_If_The_
 	req := request()
 	foundUserName, _ := domain.NewUserNameValueObject("wadus")
 	foundUser := domain.UserEntity{Name: foundUserName}
-	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: int32(1)}).Return(&foundUser, nil).Once()
+	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: 1}).Return(&foundUser, nil).Once()
 	mockedUsersRepo.On("ExistsUser", req.Context(), domain.UserEntity{Name: userName}).Return(true, nil).Once()
 
 	result := UpdateUserHandler(httptest.NewRecorder(), req, h)
@@ -201,7 +201,7 @@ func TestUpdateUserHandler_Returns_An_ErrorResult_With_An_UnexpectedError_If_Gen
 	req := request()
 	foundUserName, _ := domain.NewUserNameValueObject("wadus")
 	foundUser := domain.UserEntity{Name: foundUserName}
-	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: int32(1)}).Return(&foundUser, nil).Once()
+	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: 1}).Return(&foundUser, nil).Once()
 	mockedPassGen.On("GenerateFromPassword", "newPass").Return("", fmt.Errorf("some error")).Once()
 
 	result := UpdateUserHandler(httptest.NewRecorder(), req, h)
@@ -232,7 +232,7 @@ func TestUpdateUserHandler_Returns_An_ErrorResult_With_An_UnexpectedError_If_The
 	req := request()
 	foundUserName, _ := domain.NewUserNameValueObject("wadus")
 	foundUser := domain.UserEntity{Name: foundUserName}
-	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: int32(1)}).Return(&foundUser, nil).Once()
+	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: 1}).Return(&foundUser, nil).Once()
 	foundUser.Name = updatedUserName
 	mockedUsersRepo.On("Update", req.Context(), &foundUser).Return(nil, fmt.Errorf("some error")).Once()
 
@@ -263,10 +263,10 @@ func TestUpdateUserHandler_Updates_The_UserName(t *testing.T) {
 
 	req := request()
 	foundUserName, _ := domain.NewUserNameValueObject("wadus")
-	foundUser := domain.UserEntity{ID: int32(1), Name: foundUserName, IsAdmin: false}
-	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: int32(1)}).Return(&foundUser, nil).Once()
+	foundUser := domain.UserEntity{ID: 1, Name: foundUserName, IsAdmin: false}
+	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: 1}).Return(&foundUser, nil).Once()
 	mockedUsersRepo.On("ExistsUser", req.Context(), domain.UserEntity{Name: updatedUserName}).Return(false, nil).Once()
-	foundUser2 := domain.UserEntity{ID: int32(1), Name: foundUserName, IsAdmin: false}
+	foundUser2 := domain.UserEntity{ID: 1, Name: foundUserName, IsAdmin: false}
 	foundUser2.Name = updatedUserName
 	mockedUsersRepo.On("Update", req.Context(), &foundUser2).Return(&foundUser2, nil).Once()
 
@@ -303,8 +303,8 @@ func TestUpdateUserHandler_Updates_The_Password(t *testing.T) {
 	}
 
 	req := request()
-	foundUser := domain.UserEntity{ID: int32(1), Name: userName, IsAdmin: false}
-	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: int32(1)}).Return(&foundUser, nil).Once()
+	foundUser := domain.UserEntity{ID: 1, Name: userName, IsAdmin: false}
+	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: 1}).Return(&foundUser, nil).Once()
 	mockedPassGen.On("GenerateFromPassword", "newPass").Return("hassedPass", nil).Once()
 	mockedUsersRepo.On("Update", req.Context(), &foundUser).Return(&foundUser, nil).Once()
 
@@ -341,8 +341,8 @@ func TestUpdateUserHandler_Updates_The_IsAmin(t *testing.T) {
 	}
 
 	req := request()
-	foundUser := domain.UserEntity{ID: int32(1), Name: userName, IsAdmin: false}
-	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: int32(1)}).Return(&foundUser, nil).Once()
+	foundUser := domain.UserEntity{ID: 1, Name: userName, IsAdmin: false}
+	mockedUsersRepo.On("FindUser", req.Context(), domain.UserEntity{ID: 1}).Return(&foundUser, nil).Once()
 	foundUser.IsAdmin = true
 	mockedUsersRepo.On("Update", req.Context(), &foundUser).Return(&foundUser, nil).Once()
